requirements.txt
# Dependencias principales
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# IA - Clasificación y Regresión
scikit-learn==1.3.2
numpy==1.24.0
pandas==2.0.0

# Visión computacional - OCR
pytesseract==0.3.10
Pillow==10.0.0
opencv-python-headless==4.8.1.78

# Utilidades adicionales
python-multipart==0.0.6
pydantic==2.5.0 